pyproject.toml
[tool.poetry]
name = "standup"
version = "0.0.1"
description = "Write, publish, and keep track of your async standup reports"
authors = [
    "Samuel Dion-Girardeau <samuel.diongirardeau@gmail.com>"
]

readme = "README.md"

repository = "https://github.com/samueldg/async-standups"

keywords = ["standup"]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]


[tool.poetry.dependencies]
python = "^3.8"
click = "~=8.0"
jinja2 = "~=3.0"
pyyaml = "~=6.0"
slack_sdk = "~=3.14"

[tool.poetry.dev-dependencies]
black = "~=22.0"
flake8 = "~=4.0"
flake8-colors = "~=0.1.9"
flake8-commas = "~=2.1"
flake8-isort = "~=4.1"
flake8-quotes = "~=3.3"
isort = {version = "~=5.10", extras = ["pyproject"]}

[tool.isort]
force_single_line = true
sections = [
    "STDLIB",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]
skip_glob = [
    ".git",
    "__pycache__",
    ".vscode",
]
line_length = 99
lines_after_imports = 2

[tool.poetry.scripts]
standup = "standup.main:cli"

[build-system]
requires = ["poetry~=1.0"]
build-backend = "poetry.masonry.api"
